Move all python files to growbox module. Create run.py script.
